requirements.txt
ujson==4.0.2
xmltodict==0.12.0
aiohttp[speedups]==3.7.4.post0
discord.py==1.7.3
rich==10.6.0